database migration added using flask-migrate(alembic)

+++ resources/comments.py
@@ -1,8 +1,8 @@
 from flask import request
 from flask.views import MethodView
 from flask_smorest import abort, Blueprint
-from datetime import datetime
 from sqlalchemy.exc import SQLAlchemyError
+from flask_jwt_extended import jwt_required, get_jwt_identity
 
 from db import db
 from models import CommentsModel
@@ -17,7 +17,7 @@ class Comments(MethodView):
         comments = CommentsModel.query.all()
         return comments
     
-    
+    @jwt_required()
     @comment_blp.arguments(PlainCommentSchema)
     @comment_blp.response(201, PlainCommentSchema)
     def post(self, requested_data):
@@ -31,6 +31,7 @@ class Comments(MethodView):
             abort(500, message=f"An error occurred while inserting the comment. {e}")
 
 
+@jwt_required()
 @comment_blp.route("/comments/<int:id>")
 class CommentsById(MethodView):
     @comment_blp.response(200, PlainCommentSchema)
@@ -41,12 +42,18 @@ class CommentsById(MethodView):
     @comment_blp.arguments(PlainCommentSchema)
     @comment_blp.response(200, PlainCommentSchema)
     def put(self, requested_data, id):
+        current_user = get_jwt_identity()
+        user_id = current_user['id']
+        
         comment = CommentsModel.query.get(id)
         
         if not comment:
             abort(404, message="Comment not found")
-        comment.content = requested_data["content"]
         
+        if comment.user_id != user_id:
+            abort(403, message="You do not have permission to delete this comment.")
+        
+        comment.content = requested_data["content"]
         try:
             db.session.add(comment)
             db.session.commit()
@@ -56,7 +63,14 @@ class CommentsById(MethodView):
     
     @comment_blp.response(204)
     def delete(self, id):
+        
+        current_user = get_jwt_identity()
+        user_id = current_user['id']
+        
         comments = CommentsModel.query.get_or_404(id)
+        
+        if comments.user_id != user_id:
+            abort(403, message="You do not have permission to delete this comment.")
         db.session.delete(comments)
         db.session.commit()
         return {"message": "Comment Deleted"}, 204
